services/auth/requirements.txt
flask
prometheus-flask-exporter
redis
pyjwt
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp